Move the group exceptions to their own header file.
