+++ tiledb/sm/group/group.cc
@@ -39,6 +39,7 @@
 #include "tiledb/sm/global_state/unit_test_config.h"
 #include "tiledb/sm/group/group_details_v1.h"
 #include "tiledb/sm/group/group_details_v2.h"
+#include "tiledb/sm/group/group_internal.h"
 #include "tiledb/sm/group/group_member_v1.h"
 #include "tiledb/sm/group/group_member_v2.h"
 #include "tiledb/sm/metadata/metadata.h"
